+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- rules and debtorWorkspaceId parameters to CreditNote resource
 
 ## [0.21.0] - 2026-05-28
 ### Added